fix: Clippy suggestions

+++ ccd_lcamv06/src/command.rs
@@ -40,7 +40,7 @@ impl Command {
         let [data1, data2] = match self {
             SetIntegrationTime(t) => t.to_be_bytes(),
             SetTrigerMode(m) => [*m as u8, 0x00],
-            SetAverageTime(t) => [*t as u8, 0x00],
+            SetAverageTime(t) => [*t, 0x00],
             SetSerialBaudRate(r) => [r.to_code(), 0x00],
             _ => [0x00, 0x00],
         };
